removed deprecated game page and fixed paddle size
